+++ trunk/Marius.Html/Css/Values/CssNumber.cs
@@ -40,9 +40,9 @@ namespace Marius.Html.Css.Values
         public static readonly CssNumber Two = new CssNumber(2);
         public static readonly CssNumber Fifty = new CssNumber(50);
 
-        public double Value { get; private set; }
+        public float Value { get; private set; }
 
-        public CssNumber(double value)
+        public CssNumber(float value)
         {
             Value = value;
         }
